refactor: create a handler for rendering a gallery page
create a handler for rendering a gallery page

MODULE-1-PART-3-gallery
