+++ gallery/gallery.js
@@ -1,6 +1,7 @@
 import { setExpireTimeAfterReloading } from "../utilities/token/setToken.js";
-import { renderGalleryPage, reRenderGalleryPage } from "../utilities/render/render.js";
+import { reRenderGalleryPage } from "../utilities/render/render.js";
 import { UrlManipulationService } from "../services/url-manipulation.service.js";
+import { renderGalleryPage } from "/gallery/handler.js";
 
 const urlService = new UrlManipulationService();
 
@@ -10,7 +11,7 @@ if (!urlService.tokenExists()) {
    urlService.redirectToLoginPage();
 }
 
-await renderGalleryPage();
+await renderGalleryPage(); // handler
 
 const pages = document.getElementById('pages');
 pages.onclick = await reRenderGalleryPage();
